BGV_Bootstrapping/utils_test.py
import unittest
from utils import *


class TestUtils(unittest.TestCase):

    def test_dot_prod(self):
        result = dot_prod([1, 2, 3], [4, 5, 6])
        expected = 1 * 4 + 2 * 5 + 3 * 6
        self.assertEqual(expected, result)

    def test_bit_decomp_int(self):
        self.assertEqual([0, 1, 1], bit_decomp_int(z=6, size=3))

    def test_bit_decomp_int_size(self):
        self.assertEqual([0, 1, 1, 0, 0], bit_decomp_int(z=6, size=5))

    def test_bit_decomp_deg_0(self):
        result = bit_decomp(Polynomial([6]), 3)
        expected = [Polynomial([0]), Polynomial([1]), Polynomial([1])]
        self.assertEqual(expected, result)

    def test_bit_decomp(self):
        result = bit_decomp(Polynomial([3,4,5]), 3) #3+4x+5x^2
        e = np.array([[1,1,0], [0,0,1],[1,0,1]]).transpose()
        expected = [Polynomial(x) for x in e]
        self.assertEqual(expected, result)